Improve normalization and add a 404
Better normalization of the URLs is done now and the user is requested
to re-locate if something doesn't match up. This should also bring
better SEO. Additionally, a short 404 is now added to the mix. Support
for 404 pages has been around already but was not utilized until now
and a 500 was thrown instead on those occasions, but no more!