Centralize error handling
Handling of generic server errors should be made simple. When doing
callbacks and all kinds of crazy setups, you might want to do a simple
function (err, result) { if (handleError(err)) return; /* foobar. */ }
and providing a nice standard function allows that. Throwing in an
extra argument for the styling is also a must to make it helpful for
some custom API endpoints and other awesomeness.